feat: creating students erros validaiton

+++ src/modules/students/data/services/delete-student.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { DeleteStudent } from "../../domain/usecases";
 import { StudentRepository } from "../contracts";
+import { RegisterNotFound } from "src/modules/chore/errors";
 
 @Injectable()
 export class DeleteStudentService implements DeleteStudent {
@@ -9,6 +10,12 @@ export class DeleteStudentService implements DeleteStudent {
     ) {}
 
     async execute(studentCpf: string): Promise<void> {
+        const studentExists = await this.studentRepository.getStudentByCpf(studentCpf);
+
+        if(!studentExists) {
+            throw new RegisterNotFound('cpf')
+        } 
+
         await this.studentRepository.deleteStudent(studentCpf)
     }
 }
